refactor export management in components folder
