+++ src/scripts/components/comments/index.js
@@ -1,0 +1,2 @@
+export * from './comment';
+export * from './style';
