+++ src/scripts/components/comments/style.js
@@ -1,0 +1,1 @@
+export const commentStyles = ``;
