+++ src/scripts/components/index.js
@@ -1,0 +1,5 @@
+export * from "./carousel";
+export * from "./comments";
+export * from "./links";
+export * from "./topnav";
+export * from "./footer";
