--- src/scripts/components/slider/style.js
@@ -1,3 +1,0 @@
-const SliderStyles = ``;
-
-export default SliderStyles;
