+++ src/scripts/index.js
@@ -14,11 +14,7 @@ import "./restaurant-list";
 import "./topnav-links";
 import "./hotspot";
 import "./footer";
-import "./components/footer/index";
-import "./components/carousel/index";
-import "./components/slider/index";
-import "./components/topnav/index";
-import "./components/links/index";
+import "./components";
 import App from "./app";
 
 const app = new App({ content: document.querySelector("#main") });
